[feature]divide playNowButton in two components
