+++ truco-front/src/components/matchmaking/PlayNowButton.tsx
@@ -3,7 +3,6 @@ import Timer from './matchmakingTimer';
 import { usePusherListeners } from '../../hooks/usePusherListeners';
 import { GameContext } from '../../context/gameContext';
 import React from 'react';
-import GameStateManager from '../../gameLogic/GameStateManager';
 
 
 
@@ -20,38 +19,24 @@ const PlayNowButton = () => {
   const [gameEnded, setGameEnded] = useState<boolean>(false);
   const isSearchingRef = useRef<boolean>(false);
   const [buttonText, setButtonText] = useState<string>("Play Now");
-  const [surrenderButtonText, setSurrenderButtonText] = useState<string>("");
-  const [showSurrenderButton, setShowSurrenderButton] = useState<boolean>(false);
   const MATCHMAKING_TIME_OUT = 4000;
-  const WINNER_POINTS = 15;
-  const OPPONENT = false;
 
 
   usePusherListeners(setGameEnded)
-  const gameStateManager = new GameStateManager();
 
   useEffect(() => {
     if (opponentName == ""){
       if (isSearching){
         setButtonText("Cancel");
       }
-      else{
-        setShowSurrenderButton(false); 
-      }
-    }
-    else{
-      setSurrenderButtonText("Surrender");
-    }
-    
     
     if (gameEnded) {
+      setGameEnded(false);
+      setIsSearching(false)
       setButtonText("Play now");
-       setGameEnded(false);
-     
       
-    }
     
-  }, [isSearching, opponentName, gameEnded]);
+  }}}, [isSearching, opponentName, gameEnded]);
 
 
   const toggleMatchmaking = () => {
@@ -65,15 +50,6 @@ const PlayNowButton = () => {
     }, MATCHMAKING_TIME_OUT)
   }
 
-  const handleSurrenderButton = () => {
-    setGameEnded(true);
-    setIsSearching(false)
-    isSearchingRef.current = false
-    gameStateManager.givePoints(OPPONENT, WINNER_POINTS);
-    setSurrenderButtonText("");
-  };
-
-
   return (
     <>
       <button onClick={toggleMatchmaking}  style={ isSearching  ? { backgroundColor: "gray" } : {}}
@@ -81,16 +57,6 @@ const PlayNowButton = () => {
         <h2 className='font-medium text-2xl'>{buttonText}</h2>
       </button>
       {isSearching && !opponentName && <Timer />}
-       {showSurrenderButton}  
-        <button
-          onClick={handleSurrenderButton}
-          style={opponentName !== "" ? { backgroundColor: "#cc3333" } : {} }
-          className='w-[80%] h-[50px] bg-secondary rounded-lg flex justify-center items-center mb-4 mt-4'
-        >
-          <h2 className='font-medium text-2xl'>{surrenderButtonText}</h2>
-        </button>
-
-
 </>
 );
 };
